requirements.txt
streamlit>=1.30.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
plotly>=5.18.0
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=2.1.0
wordcloud>=1.9.3
matplotlib>=3.8.0
transformers>=4.41.0
torch>=2.2.0
